refactor the getMediaAnalyticsTracker
