+++ src/lib/actions/getMediaAnalyticsTracker/getMediaAnalyticsTracker.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 Adobe. All rights reserved.
+Copyright 2024 Adobe. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
 of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,12 +10,19 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const getMediaTrackerAPI = require("./getMediaTrackerAPI");
-const instanceManager = require("../../instanceManager/index");
+module.exports =
+  ({ instanceManager, windowObject }) =>
+  settings => {
+    const { instanceName, objectName = "Media" } = settings;
+    const instance = instanceManager.getInstance(instanceName);
 
-const windowObject = window;
+    if (!instance) {
+      throw new Error(
+        `Failed to get a Media Analytics Tracker for "${instanceName}". No matching instance was configured with this name.`
+      );
+    }
 
-module.exports = getMediaTrackerAPI({
-  instanceManager,
-  windowObject
-});
+    return instance("getMediaAnalyticsTracker", {}).then(result => {
+      windowObject[objectName] = result;
+    });
+  };
