+++ src/lib/actions/getMediaAnalyticsTracker/index.js
@@ -10,19 +10,12 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-module.exports =
-  ({ instanceManager, windowObject }) =>
-  settings => {
-    const { instanceName, objectName = "Media" } = settings;
-    const instance = instanceManager.getInstance(instanceName);
+const getMediaAnalyticsTracker = require("./getMediaAnalyticsTracker");
+const instanceManager = require("../../instanceManager/index");
 
-    if (!instance) {
-      throw new Error(
-        `Failed to send event for instance "${instanceName}". No matching instance was configured with this name.`
-      );
-    }
+const windowObject = window;
 
-    return instance("getMediaAnalyticsTracker", {}).then(result => {
-      windowObject[objectName] = result;
-    });
-  };
+module.exports = getMediaAnalyticsTracker({
+  instanceManager,
+  windowObject
+});
